Rename the "group" package to "osutil" (thanks Gustavo)
